[AIDEN] feat(A2 Tier 2 PR B): descope stranded /campaigns + type extraction — ~1,700 LOC removed
PR B of Max's Tier 2 split. PR A (#652) covered /leads + /billing
without refactor — those mock files weren't shared with KEEP
components. PR B handles the campaigns case where 6 KEEP components
(used via plasmic-init.ts + dashboard-v2/DashboardHome) imported
UI-side types and helper consts from mock-campaigns.ts.

## Type extraction (preliminary)

frontend/components/campaigns/types.ts (NEW, 56 LOC) — extracts the
UI-side shape from mock-campaigns.ts so the KEEPs can import without
depending on the mock file:

  - Campaign, MetricData, SequenceStep, CampaignStatus,
    SequenceStepStatus interfaces
  - statusStyles helper map
  - re-exports ChannelType + channelEmoji from @/data/types

Distinct from frontend/lib/api/types.ts which holds the API contract
shape (e.g. CampaignStatus enum diverges: API has 'completed', UI has
'complete' — different concerns, kept separate).

6 component imports refactored from '@/data/mock-campaigns' → './types':
CampaignCard, CampaignMetrics, CampaignSequence, CampaignChannels,
CampaignStatusBadge, SequenceStep.

## Deletions

- frontend/app/campaigns/page.tsx (stranded route, no inbound nav)
- frontend/data/mock-campaigns.ts (mock data)
- frontend/components/campaigns/{BestContentCard, RecommendationsCard,
  CampaignAllocationManager, CampaignTabs, SequenceBuilder,
  permission-mode-selector}.tsx (6 components, all zero external
  importers verified)
- 2 stale exports from components/campaigns/index.ts barrel
  (BestContentCard, RecommendationsCard)

## KEEPS (transitive deps preserved)

CampaignCard (plasmic) → CampaignStatusBadge, CampaignChannels,
  CampaignMetrics, CampaignSequence, InsightBox
CampaignSequence → SequenceStep
CampaignPriorityCard (plasmic) → PrioritySlider
CampaignMetricsPanel, CampaignPriorityPanel (plasmic) — leaf nodes
TargetingFilters (LIVE: /dashboard/campaigns/new) — leaf node

All 12 KEEP components retained. None deleted.

## Verification
  $ pnpm tsc --noEmit  → exit 0
  $ pnpm test          → 3 failed | 48 passed
    Same 3 pre-existing failures (provider-labels × 3,
    useLiveActivityFeed). Unrelated.
