+++ frontend/components/campaigns/index.ts
@@ -5,5 +5,3 @@ export { CampaignChannels } from './CampaignChannels';
 export { CampaignStatusBadge } from './CampaignStatusBadge';
 export { SequenceStep } from './SequenceStep';
 export { InsightBox } from './InsightBox';
-export { BestContentCard } from './BestContentCard';
-export { RecommendationsCard } from './RecommendationsCard';
